Change API implementation
